Script for sourcing photos from our Google Drive
From Adam Gold (UNC-CH) - this script will download photos from flood events to a specified folder and create a dataframe with a bunch of info that can be saved as an .RDS file (R object) and a CSV. Make sure to change the paths, and let me know if there are any questions! It will take awhile, mainly because Google Drive can be slow. We will need to upload the files to the Flood CamML google drive and document the dataset in our tracking sheet. I think we can randomly choose photos of non-flood events from Google Drive. Let me know when you are ready to do thi @JoeWTG.